PELT: use variable penalty for benchmark repetations
Previously, the first measurement of each by_param entry determined its
penalty, and all subsequenc measurements in it were analyzed with this
penalty. This lead to some of them coming up with a different number of
sub-states and being ignored.

Now, all by_param entries are passed to the kneedle / plateau algorithm, and
each gets its own penalty. In most cases, this leads to all entries
rporting the same number of sub-states, so far less need to be ignored.